+++ R/LearnerClustFarthestFirst.R
@@ -47,8 +47,7 @@ LearnerClustFarthestFirst = R6Class(
   private = list(
     .train = function(task) {
       pv = self$param_set$get_values(tags = "train")
-      names(pv) = chartr("_", "-", names(pv))
-      ctrl = invoke(RWeka::Weka_control, .args = pv)
+      ctrl = weka_control(pv)
       m = invoke(RWeka::FarthestFirst, x = task$data(), control = ctrl)
       if (self$save_assignments) {
         self$assignments = unname(m$class_ids + 1L)