+++ R/LearnerClustMiniBatchKMeans.R
@@ -90,7 +90,7 @@ LearnerClustMiniBatchKMeans = R6Class(
       prob = NULL
       if (self$predict_type == "prob") {
         prob = invoke(predict, self$model, newdata = data, fuzzy = TRUE)
-        colnames(prob) = seq_len(ncol(prob))
+        colnames(prob) = seq_col(prob)
       }
       PredictionClust$new(task = task, partition = partition, prob = prob)
     }