+++ R/helper.R
@@ -21,3 +21,15 @@ check_centers = function(x) {
     "`centers` must be integer or data.frame with initial cluster centers"
   }
 }
+
+row_any_na = function(x) {
+  if (!anyNA(x)) {
+    return(logical(nrow(x)))
+  }
+  rowSums(is.na(x)) > 0L
+}
+
+weka_control = function(pv) {
+  names(pv) = chartr("_", "-", names(pv))
+  invoke(RWeka::Weka_control, .args = pv)
+}