+++ NEWS.md
@@ -1,5 +1,8 @@
 # mlr3cluster (development version)
 
+* feat: Add Genie hierarchical clustering learner `clust.genie` from the genieclust package.
+* feat: Add self-organizing maps clustering learner `clust.som` from the kohonen package.
+* feat: Add robust trimmed clustering learner `clust.tclust` from the tclust package.
 * feat: Add ST-DBSCAN clustering learner `clust.stdbscan` from the stdbscan package.
 * feat: Add `clust.avg_between` measure for average between-cluster distance.
 * feat: Add `clust.avg_within` measure for average within-cluster distance.